refactor: Replace product-service with products
