--- product-service/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/productservice/command/CreateProductCommand.java
@@ -1,20 +1,0 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.productservice.command;
-
-import java.math.BigDecimal;
-
-import org.axonframework.modelling.command.TargetAggregateIdentifier;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Builder
-@Data
-public class CreateProductCommand {
-
-	@TargetAggregateIdentifier
-	private final String productId;
-	private final String title;
-	private final BigDecimal price;
-	private final Integer quantity;
-		
-}
