--- product-service/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/productservice/command/rest/ProductRestModel.java
@@ -1,14 +1,0 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.productservice.command.rest;
-
-import java.math.BigDecimal;
-
-import lombok.Data;
-
-@Data
-public class ProductRestModel {
-
-	private String title;
-	private BigDecimal price;
-	private Integer quantity;
-
-}
