--- product-service/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/productservice/core/errorhandling/ProductServiceEventsErrorHandler.java
@@ -1,16 +1,0 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.productservice.core.errorhandling;
-
-import org.axonframework.eventhandling.EventMessage;
-import org.axonframework.eventhandling.EventMessageHandler;
-import org.axonframework.eventhandling.ListenerInvocationErrorHandler;
-
-public class ProductServiceEventsErrorHandler implements ListenerInvocationErrorHandler{
-
-	@Override
-	public void onError(Exception arg0, EventMessage<?> arg1, EventMessageHandler arg2) throws Exception {
-
-		throw arg0;
-		
-	}
-
-}
