--- product-service/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/productservice/core/events/ProductCreatedEvent.java
@@ -1,15 +1,0 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.productservice.core.events;
-
-import java.math.BigDecimal;
-
-import lombok.Data;
-
-@Data
-public class ProductCreatedEvent {
-
-	private String productId;
-	private String title;
-	private BigDecimal price;
-	private Integer quantity;
-	
-}
