--- product-service/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/productservice/query/FindProductsQuery.java
@@ -1,5 +1,0 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.productservice.query;
-
-public class FindProductsQuery {
-
-}
