--- product-service/src/test/java/com/github/itsAkshayDubey/eventdrivenarchitecture/productservice/ProductServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.productservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ProductServiceApplicationTests {
-
-//	@Test
-//	void contextLoads() {
-//	}
-
-}
