+++ products/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/products/core/entity/Product.java
@@ -1,0 +1,25 @@
+package com.github.itsAkshayDubey.eventdrivenarchitecture.products.core.entity;
+
+import java.math.BigDecimal;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+import lombok.Data;
+
+@Entity
+@Table(name = "products")
+@Data
+public class Product {
+	
+	@Id
+	@Column(unique = true)
+	private String productId;
+	@Column(unique = true)
+	private String title;
+	private BigDecimal price;
+	private Integer quantity;
+
+}
