+++ products/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/products/core/entity/ProductLookupEntity.java
@@ -1,0 +1,32 @@
+package com.github.itsAkshayDubey.eventdrivenarchitecture.products.core.entity;
+
+import java.io.Serializable;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Data
+@Table(name="product_lookup")
+@AllArgsConstructor
+@NoArgsConstructor
+public class ProductLookupEntity implements Serializable {
+
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 1L;
+	
+	@Id
+	private String productId;
+	
+	@Column(unique = true)
+	private String title;
+
+}
