+++ products/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/products/core/events/ProductCreatedEvent.java
@@ -1,0 +1,15 @@
+package com.github.itsAkshayDubey.eventdrivenarchitecture.products.core.events;
+
+import java.math.BigDecimal;
+
+import lombok.Data;
+
+@Data
+public class ProductCreatedEvent {
+
+	private String productId;
+	private String title;
+	private BigDecimal price;
+	private Integer quantity;
+	
+}
