+++ products/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/products/query/FindProductsQuery.java
@@ -1,0 +1,5 @@
+package com.github.itsAkshayDubey.eventdrivenarchitecture.products.query;
+
+public class FindProductsQuery {
+
+}
