+++ products/src/test/java/com/github/itsAkshayDubey/eventdrivenarchitecture/products/ProductsTests.java
@@ -1,0 +1,13 @@
+package com.github.itsAkshayDubey.eventdrivenarchitecture.products;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ProductsTests {
+
+//	@Test
+//	void contextLoads() {
+//	}
+
+}
